+++ src/components/Channel.js
@@ -38,7 +38,10 @@ const Channel = withChatContext(
        *
        * Defaults to and accepts same props as: [LoadingIndicator](https://getstream.github.io/stream-chat-react-native/#loadingindicator)
        */
-      LoadingIndicator: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      LoadingIndicator: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * Error indicator UI component. This will be shown on the screen if channel query fails.
        *
@@ -47,7 +50,7 @@ const Channel = withChatContext(
        * */
       LoadingErrorIndicator: PropTypes.oneOfType([
         PropTypes.node,
-        PropTypes.func,
+        PropTypes.elementType,
       ]),
       /**
        * Empty state indicator UI component. This will be shown on the screen if channel has no messages.
@@ -57,7 +60,7 @@ const Channel = withChatContext(
        * */
       EmptyStateIndicator: PropTypes.oneOfType([
         PropTypes.node,
-        PropTypes.func,
+        PropTypes.elementType,
       ]),
       /**
        * Message UI component to display a message in message list.
@@ -65,13 +68,13 @@ const Channel = withChatContext(
        * Available built-in component (also accepts the same props as): [MessageSimple](https://getstream.github.io/stream-chat-react-native/#messagesimple)
        *
        * */
-      Message: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Message: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
       /**
        * Attachment UI component to display attachment in individual message.
        *
        * Available built-in component (also accepts the same props as): [Attachment](https://getstream.github.io/stream-chat-react-native/#attachment)
        * */
-      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
     };
 
     static defaultProps = {