+++ src/components/FileUploadPreview.js
@@ -31,7 +31,10 @@ export class FileUploadPreview extends React.PureComponent {
      * Custom UI component for attachment icon for type 'file' attachment.
      * Defaults to and accepts same props as: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/FileIcon.js
      */
-    AttachmentFileIcon: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+    AttachmentFileIcon: PropTypes.oneOfType([
+      PropTypes.node,
+      PropTypes.elementType,
+    ]),
   };
 
   static defaultProps = {