+++ src/components/Message.js
@@ -42,12 +42,12 @@ const Message = withKeyboardContext(
        * Message UI component to display a message in message list.
        * Avaialble from [channel context](https://getstream.github.io/stream-chat-react-native/#channelcontext)
        * */
-      Message: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Message: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
       /**
        * Attachment UI component to display attachment in individual message.
        * Avaialble from [channel context](https://getstream.github.io/stream-chat-react-native/#channelcontext)
        * */
-      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
       /**
        * Array of allowed actions on message. e.g. ['edit', 'delete', 'mute', 'flag']
        * If all the actions need to be disabled, empty array or false should be provided as value of prop.