+++ src/components/MessageList.js
@@ -82,14 +82,17 @@ const MessageList = withChannelContext(
       /** **Available from [chat context](https://getstream.github.io/stream-chat-react-native/#chatcontext)** */
       client: PropTypes.object,
       /** **Available from [channel context](https://getstream.github.io/stream-chat-react-native/#channelcontext)** */
-      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
       /**
        * Custom UI component for attachment icon for type 'file' attachment.
        * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/FileIcon.js
        */
-      AttachmentFileIcon: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      AttachmentFileIcon: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /** **Available from [channel context](https://getstream.github.io/stream-chat-react-native/#channelcontext)** */
-      Message: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Message: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
       /** **Available from [channel context](https://getstream.github.io/stream-chat-react-native/#channelcontext)** */
       messages: PropTypes.array.isRequired,
       /** **Available from [channel context](https://getstream.github.io/stream-chat-react-native/#channelcontext)** */
@@ -141,20 +144,29 @@ const MessageList = withChannelContext(
        *
        * Defaults to and accepts same props as: [TypingIndicator](https://getstream.github.io/stream-chat-react-native/#typingindicator)
        * */
-      TypingIndicator: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      TypingIndicator: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * @deprecated User DateSeperator instead.
        * Date separator UI component to render
        *
        * Defaults to and accepts same props as: [DateSeparator](https://getstream.github.io/stream-chat-react-native/#dateseparator)
        * */
-      dateSeparator: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      dateSeparator: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * Date separator UI component to render
        *
        * Defaults to and accepts same props as: [DateSeparator](https://getstream.github.io/stream-chat-react-native/#dateseparator)
        * */
-      DateSeparator: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      DateSeparator: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * @deprecated User EventIndicator instead.
        *
@@ -165,7 +177,10 @@ const MessageList = withChannelContext(
        *
        * Defaults to and accepts same props as: [EventIndicator](https://getstream.github.io/stream-chat-react-native/#eventindicator)
        * */
-      eventIndicator: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      eventIndicator: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * UI Component to display following events in messagelist
        *
@@ -174,31 +189,53 @@ const MessageList = withChannelContext(
        *
        * Defaults to and accepts same props as: [EventIndicator](https://getstream.github.io/stream-chat-react-native/#eventindicator)
        * */
-      EventIndicator: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      EventIndicator: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
 
       /** UI component for empty message list */
       EmptyStateIndicator: PropTypes.oneOfType([
         PropTypes.node,
-        PropTypes.func,
+        PropTypes.elementType,
       ]),
       /**
        * @deprecated Use HeaderComponent instead.
        *
        * UI component for header of message list.
        */
-      headerComponent: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      headerComponent: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * UI component for header of message list. By default message list doesn't have any header.
        * This is basically a [ListFooterComponent](https://facebook.github.io/react-native/docs/flatlist#listheadercomponent) of FlatList
        * used in MessageList. Its footer instead of header, since message list is inverted.
        *
        */
-      HeaderComponent: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      HeaderComponent: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
       /**
        * Style object for actionsheet (used to message actions).
        * Supported styles: https://github.com/beefe/react-native-actionsheet/blob/master/lib/styles.js
        */
       actionSheetStyles: PropTypes.object,
+      /**
+       * Besides existing (default) UX behaviour of underlying flatlist of MessageList component, if you want
+       * to attach some additional props to un derlying flatlist, you can add it to following prop.
+       *
+       * You can find list of all the available FlatList props here - https://facebook.github.io/react-native/docs/flatlist#props
+       *
+       * e.g.
+       * ```
+       * <MessageList
+       *  additionalFlatListProps={{ bounces: true, keyboardDismissMode: true }} />
+       * ```
+       */
+      additionalFlatListProps: PropTypes.object,
     };
 
     static defaultProps = {
@@ -208,6 +245,7 @@ const MessageList = withChannelContext(
       // https://github.com/facebook/react-native/blob/a7a7970e543959e9db5281914d5f132beb01db8d/Libraries/Lists/VirtualizedList.js#L466
       loadMoreThreshold: 2,
       messageGrouping: true,
+      additionalFlatListProps: {},
       dismissKeyboardOnMessageTouch: true,
       TypingIndicator,
     };
@@ -264,12 +302,14 @@ const MessageList = withChannelContext(
     insertDates = (messages) => {
       const newMessages = [];
       if (messages.length === 0) {
-        this.props.eventHistory.none.forEach((e) => {
-          newMessages.push({
-            type: 'channel.event',
-            event: e,
+        this.props.eventHistory &&
+          this.props.eventHistory.none &&
+          this.props.eventHistory.none.forEach((e) => {
+            newMessages.push({
+              type: 'channel.event',
+              event: e,
+            });
           });
-        });
 
         return newMessages;
       }
@@ -598,6 +638,7 @@ const MessageList = withChannelContext(
                 minIndexForVisible: 1,
                 autoscrollToTopThreshold: 10,
               }}
+              {...this.props.additionalFlatListProps}
             />
             {this.props.TypingIndicator && showTypingIndicator && (
               <TypingIndicatorContainer>