+++ src/components/MessageSimple/MessageContent.js
@@ -136,7 +136,7 @@ export const MessageContent = themed(
 
     static propTypes = {
       /** @see See [channel context](#channelcontext) */
-      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      Attachment: PropTypes.oneOfType([PropTypes.node, PropTypes.elementType]),
       /** enabled reactions, this is usually set by the parent component based on channel configs */
       reactionsEnabled: PropTypes.bool.isRequired,
       /** enabled replies, this is usually set by the parent component based on channel configs */
@@ -179,7 +179,10 @@ export const MessageContent = themed(
        * Custom UI component for attachment icon for type 'file' attachment.
        * Defaults to: https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/FileIcon.js
        */
-      AttachmentFileIcon: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+      AttachmentFileIcon: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.elementType,
+      ]),
     };
 
     static defaultProps = {