+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [0.5.0] 2019-12-23
+
+#### stream-chat-expo (Expo package)
+
+- Changes to add compatibility for Expo 36. Adding [netinfo](https://github.com/react-native-community/react-native-netinfo) as peer dependency - https://github.com/GetStream/stream-chat-react-native/issues/97
+- Disabling the longpress on image viewer (in Attachment) - https://github.com/GetStream/stream-chat-react-native/issues/100 to avoid freezing of UI.
+- Fixing broken threads issue
+- Support for `additionalFlatListProps` prop in `MessageList` and `ChannelList` component
+- Changing proptype for component type props to `elementType` instead of `func`
+
+#### stream-chat-react-native (Native package)
+
+- Disabling the longpress on image viewer (in Attachment) - https://github.com/GetStream/stream-chat-react-native/issues/100 to avoid freezing of UI.
+- Fixing broken threads issue
+- Support for `additionalFlatListProps` prop in `MessageList` and `ChannelList` component
+- Changing proptype for component type props to `elementType` instead of `func`
+
 ## [0.4.0] 2019-12-16
 
 - Adding support for customizing markdown styles - https://github.com/GetStream/stream-chat-react-native/pull/99/files#diff-ede54911d9164ea37e65e92f2e18cb91R56